requirements.txt
fastapi>=0.100.0
pydantic>=2.0
python-dotenv>=1.0.0
openai>=1.31.0
requests>=2.31.0
uvicorn>=0.25.0

